AnkerFix Python script
Add this file to the Ubuntu boot sequence and enjoy sane mouse inputs.

Since I'm lazy and can't be bothered to improve on it, it needs the sh package (https://github.com/amoffat/sh) and uses way too many regular expressions than are actually needed to do its job. It converts the output of xinput -list into a string, then looks in it for all instances of "USB Laser Game Mouse (lots of spaces) id=(a number)". For each instance, the script retrieves the id number, checks the properties of the corresponding input device (this is because on my laptop the Anker Gaming Mouse is recognized as three different devices - two mice and a keyboard - and the IDs change over different boot sequences!) and, if it finds a Device Accel Profile, proceeds to update its properties to sloooooooooooow it down.